Inlude development gems in docker file
I'm not sure what's changed but we're seeing an error where the development
gems are flagged as missing when we precompile the assets. This is because we
exclude them from the `bundle install`, however we shouldn't need them because
we're running `RAILS_ENV=production`.

Include development gems in the docker image to get the CI pipeline working
again; I will investigate further to try and get to the bottom of it though.