Remove privacy policy acceptance from placement request
According to data protection we shouldn't stop someone from using the service
if they haven't read the privacy notice. Our responsibility is to ensure we
bring the privacy notice to their attention and put it somewhere users can
easily find/access it.

Remove privacy policy from the acceptance bullet list and add a paragraph at
the top linking to the privacy policy. Remove the acceptance checkbox.