Default school availability and backfill existing data
We want new schools to have preset availability info text and accept in-school
bookings by default.

Add migration to default the `availability_info` text and set `experience_type`
to in-school.

Backfill existing data to set defaults when the `availability_info` has not yet
been specified by the user. Supports rolling back the `availability_info`,
however we will not be able to determine the previous `experience_type` in all
cases so this will remain as in-school. Using `update_all` without batching
should be fine as it'll result in a single SQL query to do the update.

The schema has been updated to add precisions to datetime attributes; this
appears to be Rails 7 behaviour and must have been missed during the upgrade.