Disable role checking via rolecheck feature flag
We have an existing feature flag `rolecheck` that appears to disable the role
checking for the initial sign in process.

Once signed in we appear to use the
`Schools::DFESignInAPI::RoleCheckedOrganisations` module to scope down the
available schools so that the user can only access the schools that they have
the GSE role enabled for in DFE sign in.

Going forward we want to disable role checking, so that a user will have access
to the GSE application if they have access to that school within DFE sign in
(removing the need to get a second stage of approval for using the GSE app).

Re-purpose `rolecheck` feature flag; when disabled (which it is on all
environments currently) users will not need the GSE service role in order to
access a school; they will only need to be assigned to the school in DFE sign
in.