Rework checkboxes to improve spacing
The checkboxes are styled as individual fieldsets as they map to different
underlying attributes. With the formbuilder this results in larger spacing
between checboxes than elsewhere on the website. Instead, we can group the
checkboxes in a single fieldset (using a generic key) and it will result in the
correct spacing to match other areas of the site.