Add dfe-analytics gem
Add the dfe-analytics gem and enable it when running locally/staging. Capture
all web requests and model updates as events (initially all model fields are
blacklisted; a separate PR will whitelist the ones we want).

The JSON secret for BigQuery has been added to the Azure keyvault; I had to
base64 encode it otherwise the private key was mangled.