Add task progress on-boarding wizard
We want to reconfigure the school on-boarding process into sections that can be
progressed in any order the user wants.

Add a `Wizard` that defines the flow-logic for how a user can progress through
steps in the on-boarding journey; certain steps are going to be
optional/not-applicable depending on previous answers and we need to ensure the
user ends up with a valid profile at the end of the journey.

Most of the behavioural logic has been lifted from the `CurrentStep` class and
then expanded to include additional state determinations (cannot start yet and
not applicable). This logic would be better encapsulated in the
`SchoolProfile`, however to keep the implementation cleaner for now I've opted
to duplicate parts and encapsulate it all in the new `Wizard`. That way we
aren't changing the behaviour of the `CurrentStep` class as part of this
feature and we can refactor it into the `SchoolProfile` when we decide if we
are keeping this new implementation.