Update on-boarding entry point for task progress on-boarding
When the `task_progress_on_boarding` feature is enabled we want to take the
user to the progress overview page instead of straight into the first
on-boarding step.

We also need to intercept the step submissions to return the user to the
progress overview page when they have finished a section.

I've updated the controller specs to make sure the redirects are correct when
the feature is enabled but I haven't updated the cucumber specs as it feels
like this would just be duplicating the tests; we can update them if/when we
permanently migrate to this journey.