Allow multiple reasons for a placement rejection
We currently only allow a single rejection reason for a placement request,
however we want to allow candidates to specify multiple reasons.

Add attributes/columns for each available rejection catagory. Populate new
attributes from the existing, single attribute as part of the migration.

Retain the original `rejection_category` attribute and ensure it stays up to
date (so when the categories change we pick the latest to populate this
attribute) -- assuming we don't need to revert it can later be removed.

Update analytics whitelist to include the new attributes.