Show first required step when no previous step given
In some scenarios the user will arrive to the on-boarding journey without the
initial step being defined (for example if they left half way through
on-boarding and returned manually later). In this scenario we want to take them
to the first step that is required but they have not completed (to be inline
with the existing on-boarding journey).