Fail healcheck on API timeout exception
We are seeing an occasional timeout on the API healthcheck endpoint which
raises an exception. Instead, we should be treating this as a failed
healthcheck because the API was not available.

A follow up ticket in the API repo will address the timeouts and ship before
this to avoid the service being flagged as down when its not (the timeout is a
red-herring, the API is still responsive on other endpoints).