Debug PG::CharacterNotInRepertoire error
We are seeing this error reasonably frequently; I've no idea how its happening
yet though but I expect a person/bot is searching using unexpected ASCII
characters.

Add debug logging to figure out what the searches are that cause it and attempt
to recover. Oddly Sentry reports this as part of an
`ActionView::TemplateError`, however its on the call to `save` so I think this
is a red-herring and the underlying exception was an
`ActiveRecord::StatementInvalid` which makes more sense and is replicatable in
the tests.