Increase cache TTL to 12 hours
We are seeing a reasonable number of ActionController::InvalidAuthenticityToken
errors; I'm struggling to get to the bottom of them, but I'm wondeirng if its
because our cache TTL is quite low and the CSRF tokens are expiring if a user
starts a search, goes away for a while and comes back later to keep searching.

Increase cache TTL to 12 hours to see if it has an impact on the number of CSRF
token errors we are seeing. The maximum we can set this to is 12 hours as we
store PII in the cache.