Record CSRF token errors as a metric
The CSRF error is unavoidable and we will see them happen on occasion (for
example when a candidate's session times out or if they cleared their cookies
before submitting a form). We only really care about these if they happen on a
scale that is unexpected (several per minute or hour).

To avoid spamming Slack with errors we can't resolve we will instead have
Sentry ignore them (which it does by default) and increment a Prometheus
counter that we can then alert on at a given threshold.