Add user model
Going forward we are going to be managing roles/permissions in the application
instead of via DfE sign in. As most role libraries expect a `User` model it
makes sense to add one, wrapping the DfE sign in user.

A convenience method to exchange a DfE sign in user for a `User` model is
provided (creating a new record if a matching one does not yet exist).