Add rolify gem to manage roles
We want to manage permissions within the app going forward (instead of via DfE
sign in). To do this we need a way of tracking who has access to which
resource/school.

Add `rolify` gem; this is a simple and well maintained gem that will enable us
to manage user roles on a per-resource/school basis.