Fix feature spec link matching
This was supposed to be matching the `Back` button on the page but the selector
was passed `back` (lowercase) and it was therefore matching links with
`feedback` in, of which there are now two (one in the header and one in the
footer).

Fix the selector so that it only matches the correct `Back` link and ignores
the `feedback` links.