Improve structure of closed/on-boarded schools email
We send a periodic email to our support team where we identify schools that are
now closed according to the GIAS school dataset, however they are still marked
as on-boarded with School Experience.

Currently we just send a list of the schools without any structure; instead, we
want to group the schools under various headings and format them with bullet
points (which the GOV.UK notify service will correctly format in emails).