Purge orphaned candidate records
Occasionally a contact will be removed from the CRM. This leaves us with a
candidate in School Experience that doesn't have any personally identifiable
information attached.

We currently display these contacts as 'unknown', but going forward we want to
periodically remove the orphaned candidates from our database.

Add a library to identify and remove orphaned candidates. As contact records
can be created in the CRM asynchronously we need to exclude recently created
candidates from this operation (as it may run before the async job creates the
corresponding contact record, therefore incorrectly being identified as
orphaned).