Remove confirmation requirement for new candidates
Currently we require new candidates to confirm their email before issuing the
placement request. We believe that this prevents a number of candidates from
successfully finishing their request (the email doesn't arrive and/or they
don't action it).

Add feature flag to allow us to trial skipping the confirmation step. We will
be adding other measures in order to reduce the risk of spam record creation.